Separate fetching of tweets from feed serving.
An update handler (kicked off by cron for now) fetches new tweets since the
last fetch and stores them in the datastore. The feed handler serves recent
tweets from the datastore.

Tweets are stored in the datastore using as JSON data that was received from
the Twitter REST API, to minimize lossyness.

This will enable more processing to be done on tweets (image extraction, URL
lengthening) without it being necessary to do it at fetch time (where we are
subject to App Engine URL request deadlines).