Optimize (cached) Tweet Digest response time
For a digest with 5 users, the fully cached case would require 317ms to serve
the request. 67% of that went to simplejson.loads. By adding trim_user=1, the
response JSON is much smaller, and is parsed faster. We still need to fetch the
user separately, but that means that we're only parsing that JSON once instead
of 40 times (once per status). The response time goes to 229ms.

Since we now make two Twitter API requests per user, even in the fully cached
case the caching overhead becomes noticeable. This is because python-twitter
first calls GetCachedTime (to see if the entry is valid) and then Get (if it is).
We would make the same memcache RPC for both requests (we store the cached
timestamp along with the cached value). By adding a "last request" cache, we
can avoid the second RPC for valid cache entries (true most of the time). This
makes the response time go to 190ms.

All response times were measured on the dev server (average of 4 requests)