Revert "Give a 200 response to Feedfetcher so that it can be cached."
This reverts commit 904b27d239140930c7fb8fb2f5c5c4052723838f.

The App Engine edge cache is now serving the cached respones to other clients,
which was not intentional. Per https://code.google.com/p/googleappengine/issues/detail?id=4277
it may not support the Vary header, so I'm abandoning this for now. Hopefully
the consistent 403 responses will lead to Trawler backing off in the crawl.