Use a semi-transparent bubble background
NetNewsWire started to allow the background color through (unclear why,
since there are no recent changes to the CSS stripping code in
https://github.com/Ranchero-Software/NetNewsWire/commits/main/Shared/Article%20Rendering/main.js).
However, it still forces the text color to be white or black depending
on dark vs. light mode, and this was resulting in white text on a light
gray background, which was unreadable.